Map HostedCheckoutDecryptReturnUrlRequest to HostedCheckoutDecryptReturnUrlResponse
